pyproject.toml
[tool.poetry]
name = "depmap-omics"
version = "0.1.0"
description = "CDS-omics pipelines"
authors = ["CDS-omics <qqin@broadinstitute.org>"]
readme = "README.md"
packages = [{include = "depmapomics"}, {include="mgenepy"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.1"
pandas = "^1.5.3"
taigapy = {git = "https://github.com/broadinstitute/taigapy.git", rev = "0b67741f18fe6e3c2f508ee84b06ca3bed853b82"}
seaborn = "^0.12.2"
depmap_omics_upload = {git = "https://github.com/broadinstitute/depmap_omics_upload", branch = "main"}
pyright = "1.1.293"
scipy="^1.0.0"
firecloud_dalmatian= {git = "https://github.com/pgm/dalmatian", rev = "5f37ef9"}

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^6.5.2"
pyright = "^1.1.290"
pytest = "^7.2.1"
pyright-ratchet = {git = "https://github.com/pgm/pyright-ratchet.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
